js: implemented typing status websocket events
Most of the work is done in the `typing` variable.
`typing.started()` and `typing.stopped()` send events through the
websocket connection to alert everyone else of the current typing
status.

Still not implemented: correct "<usermame> is typing..." messages. For
now, every {kind: updateTypingStatus} message is treated as a "typing
started" event, even though some of them are "typing stopped"